+++ src/engine/common/List.h
@@ -23,4 +23,4 @@ void List__Node__init(List__Node* node, void* data);
 void List__append(Arena* arena, List* list, void* data);
 void* List__get(List* list, u32 index);
 void List__remove(List* list, List__Node* node);
-void List__insort(Arena* arena, List* list, void* data, s32 (*sortCb)(void* a, void* b));
+void List__insort(Arena* arena, List* list, void* data, s8 (*sortCb)(void* a, void* b));
